requirements.txt
aiohttp==3.7.4.post0
amqp==5.0.6
aniso8601==9.0.1
APScheduler==3.7.0
ask==0.0.8
async-timeout==3.0.1
attrs==21.2.0
bcrypt==3.2.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
dnspython==2.1.0
docopt==0.6.2
email-validator==1.1.3
et-xmlfile==1.1.0
Flask==2.0.1
Flask-HTTPAuth==4.4.0
Flask-Login==0.5.0
flask-mongoengine==1.0.0
Flask-RESTful==0.3.9
Flask-WTF==0.15.1
greenlet==1.1.1
idna==3.2
iniconfig==1.1.1
itsdangerous==2.0.1
jdcal==1.4.1
Jinja2==3.0.1
kombu==5.1.0
MarkupSafe==2.0.1
mongo==0.2.0
mongoengine==0.23.1
multidict==5.1.0
numpy==1.21.2
openpyxl==2.4.11
packaging==21.0
pandas==1.3.3
pluggy==1.0.0
prompt-toolkit==3.0.20
py==1.10.0
pycparser==2.20
PyJWT==2.1.0
pymongo==3.12.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.1
records==0.5.3
redis==3.5.3
requests==2.26.0
schedule==1.1.0
six==1.16.0
SQLAlchemy==1.4.23
sync==1.0.0
tablib==3.0.0
toml==0.10.2
typing-extensions==3.10.0.2
tzlocal==2.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
WTForms==2.3.3
yarl==1.6.3
